Criação da base de dados usando EntityFramework e modelagem página "Aplicacao"

+++ AvaliacaoAsp/Code/Status.cs
@@ -5,7 +5,7 @@ using System.Web;
 
 namespace AvaliacaoAsp.Code
 {
-    public enum Status : int
+    public enum Status : byte
     {
         Bloqueado = 1,
         Desenvolvimento = 2,
